+++ library/src/main/java/com/wultra/android/digitalonboarding/VerificationService.kt
@@ -22,13 +22,14 @@ import android.content.Context
 import com.wultra.android.digitalonboarding.VerificationStatusNextStep.Value
 import com.wultra.android.digitalonboarding.log.WDOLogger
 import com.wultra.android.digitalonboarding.networking.model.ConsentApproveResponse
-import com.wultra.android.digitalonboarding.networking.model.ConsentResponse
+import com.wultra.android.digitalonboarding.networking.model.ConsentTextResponse
 import com.wultra.android.digitalonboarding.networking.CustomerOnboardingApi
 import com.wultra.android.digitalonboarding.networking.CustomerVerificationApi
 import com.wultra.android.digitalonboarding.networking.model.Document
 import com.wultra.android.digitalonboarding.networking.model.DocumentStatus
 import com.wultra.android.digitalonboarding.networking.model.DocumentSubmitResponse
 import com.wultra.android.digitalonboarding.networking.model.DocumentsStatusResponse
+import com.wultra.android.digitalonboarding.networking.model.FinishActivationResponse
 import com.wultra.android.digitalonboarding.networking.model.IdentityVerificationStatus
 import com.wultra.android.digitalonboarding.networking.model.OTPDetailResponse
 import com.wultra.android.digitalonboarding.networking.model.PresenceCheckResponse
@@ -41,12 +42,30 @@ import com.wultra.android.powerauth.networking.data.StatusResponse
 import com.wultra.android.powerauth.networking.error.ApiError
 import com.wultra.android.powerauth.networking.error.ApiErrorCode
 import io.getlime.security.powerauth.core.ActivationStatus
+import io.getlime.security.powerauth.core.Password
+import io.getlime.security.powerauth.exception.PowerAuthErrorCodes
+import io.getlime.security.powerauth.networking.response.CreateActivationResult
 import io.getlime.security.powerauth.networking.response.IActivationStatusListener
+import io.getlime.security.powerauth.networking.response.ICreateActivationListener
+import io.getlime.security.powerauth.networking.response.IValidatePasswordListener
+import io.getlime.security.powerauth.sdk.PowerAuthActivation
 import io.getlime.security.powerauth.sdk.PowerAuthSDK
 import okhttp3.OkHttpClient
 import java.time.Duration
 
 typealias VerificationResult = WDOResult<VerificationService.Success, VerificationService.Fail>
+
+/**
+ * User response for the consent.
+ */
+enum class ConsentResponse {
+    /** User approved the consent. */
+    APPROVED,
+    /** User declined the consent. */
+    DECLINED,
+    /** Consent is not required. */
+    NOT_REQUIRED
+}
 
 /**
  * Digital Onboarding Verification Service
@@ -137,7 +156,7 @@ class VerificationService(
 
                 when (nextStep.value) {
                     Value.INTRO -> {
-                        markCompleted(VerificationStateIntroData, callback)
+                        markCompleted(VerificationStateIntroData(nextStep.consentRequired), callback)
                     }
                     Value.DOCUMENT_SCAN -> {
                         WDOLogger.d("Asking for a document status.")
@@ -199,6 +218,9 @@ class VerificationService(
                     Value.OTP -> {
                         markCompleted(VerificationStateOtpData(null), callback)
                     }
+                    Value.ACTIVATION_FINISH -> {
+                        markCompleted(VerificationStateActivationFinishData, callback)
+                    }
                     Value.FAILED -> {
                         markCompleted(VerificationStateFailedData, callback)
                     }
@@ -224,66 +246,74 @@ class VerificationService(
      *
      * @param callback Callback with the result.
      */
-    fun consentGet(callback: (VerificationResult) -> Unit) {
-
+    fun getConsent(callback: (WDOResult<String, Fail>) -> Unit) {
         val processId = guardProcessId(callback) ?: return
-
         api.getConsentText(
             processId,
-            object : IApiCallResponseListener<ConsentResponse> {
-                override fun onSuccess(result: ConsentResponse) {
-                    WDOLogger.i("consentGet success")
-                    markCompleted(
-                        VerificationStateConsentData(result.responseObject.consentText),
-                        callback,
-                    )
+            object : IApiCallResponseListener<ConsentTextResponse> {
+                override fun onSuccess(result: ConsentTextResponse) {
+                    WDOLogger.i("getConsent success")
+                    callback(WDOResult.success(result.responseObject.consentText))
                 }
 
                 override fun onFailure(error: ApiError) {
-                    WDOLogger.e("consentGet failed : ${error.e}")
-                    markCompleted(error, callback)
+                    WDOLogger.e("getConsent failed : ${error.e}")
+                    callback(WDOResult.failure(Fail(reason = error)))
                 }
             },
         )
     }
 
     /**
-     * Approves the consent for this process and starts the activation.
+     * Starts the verification process with the user consent.
      *
+     * @param consentApprovedByUser User response for the consent.
      * @param callback Callback with the result.
      */
-    fun consentApprove(callback: (VerificationResult) -> Unit) {
-
+    fun start(consentApprovedByUser: ConsentResponse, callback: (VerificationResult) -> Unit) {
         val processId = guardProcessId(callback) ?: return
 
-        api.resolveConsent(
-            processId,
-            true,
-            object : IApiCallResponseListener<ConsentApproveResponse> {
-                override fun onSuccess(result: ConsentApproveResponse) {
-                    WDOLogger.i("consentApprove success - starting the process")
-                    api.start(
-                        processId,
-                        object : IApiCallResponseListener<StatusResponse> {
-                            override fun onSuccess(result: StatusResponse) {
-                                WDOLogger.i("Process start success")
-                                markCompleted(VerificationStateDocumentsToScanSelectData, callback)
-                            }
-
-                            override fun onFailure(error: ApiError) {
-                                WDOLogger.e("start failed : ${error.e}")
-                                markCompleted(error, callback)
-                            }
-                        },
-                    )
-                }
-
-                override fun onFailure(error: ApiError) {
-                    WDOLogger.e("consentApprove failed : ${error.e}")
-                    markCompleted(error, callback)
-                }
-            },
-        )
+        when (consentApprovedByUser) {
+            ConsentResponse.APPROVED -> {
+                WDOLogger.i("User approved consent - resolving on the server")
+                api.resolveConsent(
+                    processId,
+                    true,
+                    object : IApiCallResponseListener<ConsentApproveResponse> {
+                        override fun onSuccess(result: ConsentApproveResponse) {
+                            WDOLogger.i("Consent granted - starting the verification process.")
+                            startProcess(processId, callback)
+                        }
+                        override fun onFailure(error: ApiError) {
+                            WDOLogger.e("Consent approve failed : ${error.e}")
+                            markCompleted(error, callback)
+                        }
+                    },
+                )
+            }
+            ConsentResponse.DECLINED -> {
+                WDOLogger.i("User declined consent - returning to intro state")
+                api.resolveConsent(
+                    processId,
+                    false,
+                    object : IApiCallResponseListener<ConsentApproveResponse> {
+                        override fun onSuccess(result: ConsentApproveResponse) {
+                            WDOLogger.i("Consent declined success.")
+                            val consentRequired = lastStatus?.responseObject?.consentRequired ?: true
+                            markCompleted(VerificationStateIntroData(consentRequired), callback)
+                        }
+                        override fun onFailure(error: ApiError) {
+                            WDOLogger.e("Consent decline failed : ${error.e}")
+                            markCompleted(error, callback)
+                        }
+                    },
+                )
+            }
+            ConsentResponse.NOT_REQUIRED -> {
+                WDOLogger.i("Consent not required - start verification process immediately")
+                startProcess(processId, callback)
+            }
+        }
     }
 
     /**
@@ -433,7 +463,7 @@ class VerificationService(
             object : IApiCallResponseListener<StatusResponse> {
                 override fun onSuccess(result: StatusResponse) {
                     WDOLogger.i("restartVerification success")
-                    markCompleted(VerificationStateIntroData, callback)
+                    status(callback)
                 }
 
                 override fun onFailure(error: ApiError) {
@@ -539,11 +569,152 @@ class VerificationService(
     }
 
     /**
+     * Finishes verification by creating a new PowerAuth activation on given `newPowerAuthInstance`.
+     *
+     * Needs to be called when `ACTIVATION_FINISH` next step is returned from the `status()` call.
+     *
+     * The method verifies that the provided `password` is the same as used in the original activation
+     * (if `validatePassword` is set to `true`), then it calls the server API to finish
+     * the verification and obtain the activation code for the new activation. Finally, it creates
+     * a new activation on the `newPowerAuthInstance` using the obtained activation code and persists it
+     * with the provided `newPassword`.
+     *
+     * After successful completion, the original PowerAuth instance becomes invalid (`REMOVED` state) and cannot be used anymore.
+     *
+     * @param newPowerAuthInstance PowerAuth instance where to create new activation. This instance must not have an existing activation.
+     * @param newActivationName Name of the new activation to be created on `newPowerAuthInstance`.
+     * @param newPassword Password to protect the new activation. In case `validatePassword` is `true`, this password must match the password of the original activation.
+     * @param validatePassword If set to `true`, the method verifies that the provided `newPassword` matches the password of the original activation.
+     * @param userIdentification Optional user identification object to be sent to the server during the finish activation process.
+     */
+    fun finishActivation(
+        newPowerAuthInstance: PowerAuthSDK,
+        newActivationName: String,
+        newPassword: Password,
+        validatePassword: Boolean,
+        userIdentification: Any?,
+        callback: (VerificationResult) -> Unit
+    ) {
+
+        val processId = guardProcessId(callback) ?: return
+
+        // Validate password if required
+        validatePasswordIfRequired(validatePassword, newPassword) { error ->
+
+            if (error != null) {
+                // Password validation failed
+                WDOLogger.e("finishActivation - password validation failed : ${error.message}")
+                markCompleted(ApiError(error), callback)
+                return@validatePasswordIfRequired
+            }
+
+            if (!newPowerAuthInstance.canStartActivation()) {
+                WDOLogger.e("finishActivation - cannot activate, the `newPowerAuthInstance` is not in a state that allows it")
+                markCompleted(ApiError(Exception("Cannot start activation on given PowerAuth object.")), callback)
+                return@validatePasswordIfRequired
+            }
+
+            // Proceed with finish activation API call
+            api.finishActivation(
+                processId,
+                userIdentification,
+                object: IApiCallResponseListener<FinishActivationResponse> {
+
+                    override fun onSuccess(result: FinishActivationResponse) { // onSuccess:finishActivation
+
+                        WDOLogger.i("finishActivation success")
+
+                        // Create new activation on the new PowerAuth instance with obtained activation code
+                        val activation = PowerAuthActivation.Builder.activation(
+                            result.responseObject.activationCode,
+                            newActivationName
+                        )
+
+                        // Helper to clear the created activation in case of failure
+                        val clearPaInstanceIfNeeded = {
+                            if (!newPowerAuthInstance.canStartActivation()) {
+                                newPowerAuthInstance.removeActivationLocal(appContext)
+                            }
+                        }
+
+                        // create activation
+                        newPowerAuthInstance.createActivation(
+                            activation.build(),
+                            object : ICreateActivationListener {
+
+                                override fun onActivationCreateSucceed(result: CreateActivationResult) {
+                                    // New activation created, now persist it with the provided password
+                                    val persistResult = newPowerAuthInstance.persistActivationWithPassword(appContext, newPassword)
+                                    if (persistResult == PowerAuthErrorCodes.SUCCEED) {
+                                        // New activation persisted
+                                        markCompleted(VerificationStateSuccessData, callback)
+                                    } else {
+                                        // Failed to persist new activation, clean up
+                                        clearPaInstanceIfNeeded()
+                                        WDOLogger.e("Failed to persist PowerAuth activation. Code: $persistResult")
+                                        markCompleted(Fail(ApiError(Exception("Failed to persist PowerAuth activation. Code: $persistResult"))), callback)
+                                    }
+                                }
+
+                                override fun onActivationCreateFailed(t: Throwable) {
+                                    // Failed to create new activation, clean up
+                                    clearPaInstanceIfNeeded()
+                                    WDOLogger.e("finishActivation failed - failed to create activation : $t")
+                                    markCompleted(Fail(ApiError(t)), callback)
+                                }
+                            }
+                        )
+                    }
+
+                    override fun onFailure(error: ApiError) { // onFailure:finishActivation
+                        // Finish activation API call failed
+                        WDOLogger.e("finishActivation failed : ${error.e}")
+                        markCompleted(error, callback)
+                    }
+                }
+            )
+        }
+    }
+
+    /**
+     * Validates password if required. If not required, calls callback immediately.
+     *
+     * @param required Whether the password validation is required.
+     * @param password Password to validate.
+     * @param callback Callback with the error if any.
+     */
+    private fun validatePasswordIfRequired(
+        required: Boolean,
+        password: Password,
+        callback: (Throwable?) -> Unit,
+    ) {
+        if (!required) {
+            // Password validation not required
+            callback(null)
+        } else {
+            powerAuthSDK.validatePassword(
+                appContext,
+                password,
+                object : IValidatePasswordListener {
+                    override fun onPasswordValid() {
+                        callback(null)
+                    }
+
+                    override fun onPasswordValidationFailed(t: Throwable) {
+                        WDOLogger.i("Password validation failed.")
+                        callback(t)
+                    }
+                }
+            )
+        }
+    }
+
+    /**
      * Demo endpoint available only in Wultra Demo systems.
      *
      * @param callback Callback with the result.
      */
-    fun getOTP(callback: (WDOResult<String, Fail>) -> Unit) {
+    internal fun getOTP(callback: (WDOResult<String, Fail>) -> Unit) {
 
         val processId = guardProcessId(callback) ?: return
 
@@ -599,6 +770,23 @@ class VerificationService(
     object OTPFailedException: Exception("OTP failed to verify.")
 
     // Private helper methods
+
+    // Starts the verification process on backend.
+    private fun startProcess(processId: String, callback: (VerificationResult) -> Unit) {
+        api.start(
+            processId,
+            object : IApiCallResponseListener<StatusResponse> {
+                override fun onSuccess(result: StatusResponse) {
+                    WDOLogger.i("Verification process start success")
+                    markCompleted(VerificationStateDocumentsToScanSelectData, callback)
+                }
+                override fun onFailure(error: ApiError) {
+                    WDOLogger.e("Verification process start failed : ${error.e}")
+                    markCompleted(error, callback)
+                }
+            },
+        )
+    }
 
     private fun <T>guardProcessId(callback: (WDOResult<T, Fail>) -> Unit): String? {