+++ library/src/main/java/com/wultra/android/digitalonboarding/networking/model/ConfigurationObjects.kt
@@ -19,7 +19,6 @@ package com.wultra.android.digitalonboarding.networking.model
 import com.google.gson.annotations.SerializedName
 import com.wultra.android.powerauth.networking.data.ObjectRequest
 import com.wultra.android.powerauth.networking.data.ObjectResponse
-import com.wultra.android.powerauth.networking.data.StatusResponse.Status
 
 /**
  * REQUEST
@@ -40,7 +39,9 @@ class ConfigurationResponse(
     responseObject: ConfigurationResponseData,
     status: Status
 ): ObjectResponse<ConfigurationResponseData>(responseObject, status)
-class ConfigurationResponseData(
+
+/** Configuration response */
+data class ConfigurationResponseData(
     /** Is the onboarding process enabled */
     @SerializedName("enabled") val enabled: Boolean,
     /** Is OTP required for the first part - identification/activation. */
@@ -48,21 +49,29 @@ class ConfigurationResponseData(
     /** Is OTP required for the second part - identity verification. */
     @SerializedName("otpForIdentityVerification") val otpForIdentityVerification: Boolean,
     /** Documents required for identity verification. */
-    @SerializedName("documents") val documents: ConfigurationDocumentsData
+    @SerializedName("documents") val documents: ConfigurationDocuments
 )
 
-class ConfigurationDocumentsData(
-    /** Number of required documents */
+/** Documents required for identity verification */
+data class ConfigurationDocuments(
+    /** Number of total required documents */
+    @SerializedName("totalRequiredDocumentsCount") val totalRequiredDocumentsCount: Int,
+    /** Groups of documents */
+    @SerializedName("groups") val groups: List<ConfigurationDocumentGroup>
+)
+
+/** Configuration for a document */
+class ConfigurationDocumentGroup(
+    /** Number of required documents in the group */
     @SerializedName("requiredDocumentsCount") val requiredDocumentsCount: Int,
-    /** List of documents */
-    @SerializedName("items") val items: List<ConfigurationDocumentData>
+    /** Documents in the group */
+    @SerializedName("items") val items: List<ConfigurationDocument>
 )
 
-class ConfigurationDocumentData(
+/** Group of documents in the configuration */
+data class ConfigurationDocument(
     /** Type of the document */
     @SerializedName("type") val type: String,
-    /** Is the document mandatory? */
-    @SerializedName("mandatory") val mandatory: Boolean,
     /** Number of sides the document has */
     @SerializedName("sideCount") val sideCount: Int
 )